ANS update 2022-06-21 [update].

+++ src/main/java/com/gemiso/zodiac/app/appInterface/takerCueFindAllDTO/TakerCueSheetSpareDTO.java
@@ -62,7 +62,7 @@ public class TakerCueSheetSpareDTO {
     @XmlElement(name = "rptr_nm")
     private String rptrNm;
     @XmlElement(name = "dept_cd")
-    private Integer deptCd;
+    private Long deptCd;
     @XmlElement(name = "dept_nm")
     private String deptNm;
     @XmlElement(name = "artcl_reqd_sec")
