+++ src/main/java/com/gemiso/zodiac/app/article/Article.java
@@ -244,7 +244,7 @@ public class Article extends BaseEntity {
     private Integer videoTime;
 
     @Column(name = "dept_cd", length = 50)
-    private Integer deptCd;
+    private Long deptCd;
 
     //@Basic(fetch = FetchType.LAZY)
     @Formula("(select a.name from tb_depts a where a.id = dept_cd)")
