+++ src/main/java/com/gemiso/zodiac/app/article/ArticleRepositoryCustorm.java
@@ -11,14 +11,14 @@ public interface ArticleRepositoryCustorm {
 
     Page<Article> findByArticleList(Date sdate, Date edate, Date rcvDt, String rptrId, String inputrId, String brdcPgmId,
                         String artclDivCd, String artclTypCd, String searchDivCd, String searchWord,
-                        List<String> apprvDivCdList, Integer deptCd, String artclCateCd, String artclTypDtlCd,
+                        List<String> apprvDivCdList, Long deptCd, String artclCateCd, String artclTypDtlCd,
                         String delYn, Long artclId, String copyYn, Long orgArtclId, Pageable pageable);
 
     Page<Article> findByArticleIssue(Date sdate, Date edate, String issuKwd, String artclDivCd, String artclTypCd,
-                                     String artclTypDtlCd, String artclCateCd, Integer deptCd, String inputrId,
+                                     String artclTypDtlCd, String artclCateCd, Long deptCd, String inputrId,
                                      String brdcPgmId, Long orgArtclId, String delYn, String searchDivCd,
                                      String searchWord, List<String> apprvDivCdList, Pageable pageable);
 
     Page<Article> findByArticleCue(Date sdate, Date edate, String searchWord, Long cueId, String brdcPgmId, String artclTypCd,
-                                   String artclTypDtlCd, String copyYn, Integer deptCd, Long orgArtclId, Pageable pageable);
+                                   String artclTypDtlCd, String copyYn, Long deptCd, Long orgArtclId, Pageable pageable);
 }
