+++ src/main/java/com/gemiso/zodiac/app/cueSheet/dto/CueSheetUpdateDTO.java
@@ -63,7 +63,7 @@ public class CueSheetUpdateDTO {
     private String brdcRunTime;
     private Integer cueVer;
     private Integer cueOderVer;
-    private Integer deptCd;
+    private Long deptCd;
     //private String deptNm;
     private BaseProgramSimpleDTO baseProgram;
     private ProgramSimpleDTO program;
